prevent loading directories missing "game.json"
I modified this as such so that the server would stop trying to load my .git and node_modules folders as games. You could check for any other required file, but game.json was/is required in the root game folder, and it is hard coded in js-game-server, so it's easy to check for.